pyproject.toml
[tool.poetry]
name = "instructed-retriever"
version = "0.1.0"
description = "A retrieval agent that propagates system specifications through every stage of the search pipeline"
readme = "README.md"
packages = [{ include = "instructed_retriever", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
httpx = ">=0.27.2"
pydantic = ">=2.10.3"
pydantic-settings = ">=2.6.1"
dspy = ">=3.1.2,<4.0.0"
litellm = ">=1.68.2,<1.76.0"
mlflow-skinny = ">=3.10.1,<4.0.0"
databricks-langchain = "^0.11.0"
databricks-sdk = ">=0.20.0"
pyyaml = ">=5.1"
presidio-analyzer = "^2.2.361"
presidio-anonymizer = "^2.2.361"
usaddress = "^0.5.16"
en-core-web-md = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl" }
requests = ">=2.31.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
ruff = "^0.11.0"
mypy = "^1.11.2"
pre-commit = "^4.0.0"
ipykernel = "^7.1.0"
jupyter = "^1.1.1"
notebook = "^7.5.2"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by ruff format
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "httpx"]

[tool.mypy]
allow_incomplete_defs = false
allow_untyped_defs = false
ignore_missing_imports = false
no_implicit_optional = true
implicit_reexport = false
local_partial_types = true
pretty = true
show_error_codes = true
show_traceback = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["dspy.*", "databricks_langchain", "databricks.*", "usaddress"]
ignore_missing_imports = true
